Restructuring and various improvements
