+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TFLiteHandler.java
@@ -1,0 +1,232 @@
+package org.firstinspires.ftc.teamcode;
+
+import com.qualcomm.robotcore.util.ElapsedTime;
+
+import org.firstinspires.ftc.robotcore.external.ClassFactory;
+import org.firstinspires.ftc.robotcore.external.navigation.VuforiaLocalizer;
+import org.firstinspires.ftc.robotcore.external.tfod.Recognition;
+import org.firstinspires.ftc.robotcore.external.tfod.TFObjectDetector;
+import com.qualcomm.robotcore.hardware.HardwareMap;
+
+import java.util.List;
+
+public class TFLiteHandler extends HardwareDefinitions{
+
+
+    private static final String TFOD_MODEL_ASSET = "RoverRuckus.tflite";
+    private static final String LABEL_GOLD_MINERAL = "Gold Mineral";
+    private static final String LABEL_SILVER_MINERAL = "Silver Mineral";
+
+    private static final String VUFORIA_KEY = "AR1NGjD/////AAABmRNHhw4urkcYu6OsCz4GxO9HaxexcrZrSNGBfCYsc8miWAyyHlu53AsvQ0AMdhXKpFuLLm0Dej3xk4agW4J4tOXGu+hPnigkbDyr5HhVrGXPGxFyNCpJUHx+Sr6UMygVYr5b+z78sdhUeN2o4KBHClV+VzRnAuG0h4GiWh+58fPYhqIIRboPe41XAbmNWwCIqAG+1y5XXaENN0jq99vO4e4GgzYzQdAQtK4Jrq4pkIZev+fI5K2B500kIkiVv3YrnC1JkQNIfibntc+98DKcN7hbJ3TWJmHndB9vesnlzPnDEJ/q9j+V+w82/icXhZ58Jcu+QMu/iuo7eEZeCLQ8S5BqotKIbxP3mCW31jh93Btc"; //put vuforia key in quotes
+
+    private VuforiaLocalizer vuforia;
+
+    private TFObjectDetector tfod;
+
+    ElapsedTime timer = new ElapsedTime();
+
+    int goldPosition = 0; //1 is left, 2 is center, 3 is right
+
+
+    public void initVuforia() {
+        /*
+         * Configure Vuforia by creating a Parameter object, and passing it to the Vuforia engine.
+         */
+        VuforiaLocalizer.Parameters parameters = new VuforiaLocalizer.Parameters();
+
+        parameters.vuforiaLicenseKey = VUFORIA_KEY;
+        parameters.cameraDirection = VuforiaLocalizer.CameraDirection.BACK;
+
+        //  Instantiate the Vuforia engine
+        vuforia = ClassFactory.getInstance().createVuforia(parameters);
+
+        // Loading trackables is not necessary for the Tensor Flow Object Detection engine.
+    }
+
+    /**
+     * Initialize the Tensor Flow Object Detection engine.
+     */
+    public void initTfod() {
+        int tfodMonitorViewId = hardwareMap.appContext.getResources().getIdentifier(
+                "tfodMonitorViewId", "id", hardwareMap.appContext.getPackageName());
+        TFObjectDetector.Parameters tfodParameters = new TFObjectDetector.Parameters(tfodMonitorViewId);
+        tfod = ClassFactory.getInstance().createTFObjectDetector(tfodParameters, vuforia);
+        tfod.loadModelFromAsset(TFOD_MODEL_ASSET, LABEL_GOLD_MINERAL, LABEL_SILVER_MINERAL);
+    }
+
+    public void initTFodAndVuforia() {
+        initVuforia();
+
+        if (ClassFactory.getInstance().canCreateTFObjectDetector()) {
+            initTfod();
+            telemetry.addData("Initialized TensorFlow and Vuforia", "");
+        } else {
+            telemetry.addData("OOPS!!", "Your device can't work with TFOD");
+        }
+    }
+
+    public void detectGold_inferRight(double maxTimeS){ //looks at left and center mineral
+
+        timer.reset();
+
+        if (opModeIsActive()) {
+            /** Start Tensor Flow Object Detection. */
+            if (tfod != null) {
+                tfod.activate();
+            }
+
+            while (opModeIsActive() && timer.seconds() < maxTimeS) {
+                if (tfod != null) {
+                    List<Recognition> updatedRecognitions = tfod.getUpdatedRecognitions();
+                    if (updatedRecognitions != null) {
+                        telemetry.addData("# Object Detected", updatedRecognitions.size());
+                        if (updatedRecognitions.size() == 2) {
+                            int goldMineralX = -1;
+                            int silverMineralX = -1;
+                            for (Recognition recognition : updatedRecognitions) {
+                                if (recognition.getLabel().equals(LABEL_GOLD_MINERAL)) {
+                                    goldMineralX = (int) recognition.getLeft();
+                                } else if (silverMineralX == -1) {
+                                    silverMineralX = (int) recognition.getLeft();
+                                }
+                            }
+                            if (goldMineralX == -1) {
+                                //telemetry.addData("Mineral Position:", "Left");
+                                telemetry.addData("Mineral Position:", "Right");
+                                //goldPosition = 1;
+                                goldPosition = 3;
+                            } else if (goldMineralX != -1){
+                                if(goldMineralX < silverMineralX){
+                                    //telemetry.addData("Mineral Position:", "Center");
+                                    telemetry.addData("Mineral Position:", "Left");
+                                    //goldPosition = 2;
+                                    goldPosition = 1;
+                                } else if(goldMineralX > silverMineralX){
+                                    //telemetry.addData("Mineral Position:", "Right");
+                                    telemetry.addData("Mineral Position:", "Center");
+                                    //goldPosition = 3;
+                                    goldPosition = 2;
+                                } else {
+                                    telemetry.addData("Mineral Position:", "Confused");
+                                }
+                            }
+                        }
+                        telemetry.update();
+                    }
+                }
+            }
+        }
+        if (tfod != null) {
+            tfod.shutdown();
+        }
+    }
+
+    public void detectGold_inferLeft(double maxTimeS){
+
+        timer.reset();
+
+        if (opModeIsActive()) {
+            /** Start Tensor Flow Object Detection. */
+            if (tfod != null) {
+                tfod.activate();
+            }
+
+            while (opModeIsActive() && timer.seconds() < maxTimeS) {
+                if (tfod != null) {
+                    List<Recognition> updatedRecognitions = tfod.getUpdatedRecognitions();
+                    if (updatedRecognitions != null) {
+                        telemetry.addData("# Object Detected", updatedRecognitions.size());
+                        if (updatedRecognitions.size() == 2) {
+                            int goldMineralX = -1;
+                            int silverMineralX = -1;
+                            for (Recognition recognition : updatedRecognitions) {
+                                if (recognition.getLabel().equals(LABEL_GOLD_MINERAL)) {
+                                    goldMineralX = (int) recognition.getLeft();
+                                } else if (silverMineralX == -1) {
+                                    silverMineralX = (int) recognition.getLeft();
+                                }
+                            }
+                            if (goldMineralX == -1) {
+                                telemetry.addData("Mineral Position:", "Left");
+                                goldPosition = 1;
+                            } else if (goldMineralX != -1){
+                                if(goldMineralX > silverMineralX){
+                                    telemetry.addData("Mineral Position:", "Center");
+                                    goldPosition = 2;
+                                } else if(goldMineralX < silverMineralX){
+                                    telemetry.addData("Mineral Position:", "Right");
+                                    goldPosition = 3;
+                                } else {
+                                    telemetry.addData("Mineral Position:", "Confused");
+                                }
+                            }
+                        }
+                        telemetry.update();
+                    }
+                }
+            }
+        }
+        if (tfod != null) {
+            tfod.shutdown();
+        }
+
+    }
+
+    public void detectGold_inferNone(double maxTimeS){
+
+        timer.reset();
+
+        if (opModeIsActive() && timer.seconds() < maxTimeS) {
+            /** Activate Tensor Flow Object Detection. */
+            if (tfod != null) {
+                tfod.activate();
+            }
+
+            while (opModeIsActive()) {
+                if (tfod != null) {
+                    // getUpdatedRecognitions() will return null if no new information is available since
+                    // the last time that call was made.
+                    List<Recognition> updatedRecognitions = tfod.getUpdatedRecognitions();
+                    if (updatedRecognitions != null) {
+                        telemetry.addData("# Object Detected", updatedRecognitions.size());
+                        if (updatedRecognitions.size() == 3) {
+                            int goldMineralX = -1;
+                            int silverMineral1X = -1;
+                            int silverMineral2X = -1;
+                            for (Recognition recognition : updatedRecognitions) {
+                                if (recognition.getLabel().equals(LABEL_GOLD_MINERAL)) {
+                                    goldMineralX = (int) recognition.getLeft();
+                                } else if (silverMineral1X == -1) {
+                                    silverMineral1X = (int) recognition.getLeft();
+                                } else {
+                                    silverMineral2X = (int) recognition.getLeft();
+                                }
+                            }
+                            if (goldMineralX != -1 && silverMineral1X != -1 && silverMineral2X != -1) {
+                                if (goldMineralX < silverMineral1X && goldMineralX < silverMineral2X) {
+                                    telemetry.addData("Gold Mineral Position", "Left");
+                                    goldPosition = 1;
+                                } else if (goldMineralX > silverMineral1X && goldMineralX > silverMineral2X) {
+                                    telemetry.addData("Gold Mineral Position", "Right");
+                                    goldPosition = 3;
+                                } else {
+                                    telemetry.addData("Gold Mineral Position", "Center");
+                                    goldPosition = 2;
+                                }
+                            }
+                        }
+                        telemetry.update();
+                    }
+                }
+            }
+        }
+        if (tfod != null) {
+            tfod.shutdown();
+        }
+    }
+
+    public int getMineralPosition(){
+
+        return goldPosition;
+    }
+}
