--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/BasicVision.java
@@ -1,39 +1,0 @@
-package org.firstinspires.ftc.teamcode.autonomous;
-
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-import com.qualcomm.robotcore.eventloop.opmode.Disabled;
-import com.qualcomm.robotcore.hardware.DcMotor;
-
-import org.firstinspires.ftc.teamcode.HardwareDefinitions;
-
-@Disabled
-@Autonomous(name="BasicVision", group="Autonomous")
-public class BasicVision extends HardwareDefinitions {
-
-    @Override
-    public void runOpMode() {
-
-        //INITIALIZATION PERIOD
-        init(hardwareMap);
-
-        setDriveEncoderMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-        setDriveEncoderMode(DcMotor.RunMode.RUN_USING_ENCODER);
-        landerMotor1.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-
-        telemetry.addData("Robot is initialized", "");
-        telemetry.update();
-
-
-        // Wait for the game to start (driver presses PLAY)
-        waitForStart();
-
-        //START PERIOD
-
-       // dropFromLander();
-
-
-
-
-
-    }
-}
