+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/LinearCraterSide.java
@@ -3,11 +3,11 @@ package org.firstinspires.ftc.teamcode.autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.hardware.DcMotor;
 
-import org.firstinspires.ftc.teamcode.HardwareDefinitions;
+
 
 
 @Autonomous(name="LinearCraterSide", group="Autonomous")
-public class LinearCraterSide extends HardwareDefinitions {
+public class LinearCraterSide extends AutonomousDefinitions {
 
     @Override
     public void runOpMode() {
