+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/LinearDepotSide.java
@@ -3,11 +3,10 @@ package org.firstinspires.ftc.teamcode.autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.hardware.DcMotor;
 
-import org.firstinspires.ftc.teamcode.HardwareDefinitions;
 
 
 @Autonomous(name="LinearDepotSide", group="Autonomous")
-public class LinearDepotSide extends HardwareDefinitions {
+public class LinearDepotSide extends AutonomousDefinitions {
 
 
     @Override
