+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/LinearDepotSide2.java
@@ -3,11 +3,9 @@ package org.firstinspires.ftc.teamcode.autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.hardware.DcMotor;
 
-import org.firstinspires.ftc.teamcode.HardwareDefinitions;
-
 
 @Autonomous(name="LinearDepotSide2", group="Autonomous")
-public class LinearDepotSide2 extends HardwareDefinitions {
+public class LinearDepotSide2 extends AutonomousDefinitions {
 
 
     @Override
@@ -15,6 +13,7 @@ public class LinearDepotSide2 extends HardwareDefinitions {
 
         //INITIALIZATION PERIOD
         init(hardwareMap);
+        initIMU(hardwareMap);
 
         setDriveEncoderMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
         setDriveEncoderMode(DcMotor.RunMode.RUN_USING_ENCODER);
@@ -30,7 +29,15 @@ public class LinearDepotSide2 extends HardwareDefinitions {
         markerDropperOuter.setPosition(markerDropperOuterHold);
         markerDropperInner.setPosition(markerDropperInnerHold);
 
-        dropFromLander();
+
+        gyroTurn(0.3, 90, 5);
+
+        sleep(3000);
+
+        gyroTurn(0.3, -90, 5);
+
+
+        //dropFromLander();
 
         //START PERIOD
