+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/teleop/DriverControlledTest.java
@@ -4,7 +4,10 @@ import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.util.Range;
 
-import org.firstinspires.ftc.teamcode.HardwareDefinitions;
+import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
+import org.firstinspires.ftc.robotcore.external.navigation.AxesOrder;
+import org.firstinspires.ftc.robotcore.external.navigation.AxesReference;
+import org.firstinspires.ftc.teamcode.autonomous.AutonomousDefinitions;
 
 //import com.qualcomm.robotcore.hardware.CRServo;
 //import com.qualcomm.robotcore.hardware.GyroSensor;
@@ -12,8 +15,8 @@ import org.firstinspires.ftc.teamcode.HardwareDefinitions;
 
 //See Google Drive for TODO
 
-@TeleOp(name="DriverControlled2", group="Test")
-public class DriverControlled2 extends HardwareDefinitions{
+@TeleOp(name="DriverControlledTest", group="Test")
+public class DriverControlledTest extends AutonomousDefinitions{
 
     //LIFT SERVO VARIABLES
     boolean opener1Changed = false;
@@ -30,6 +33,7 @@ public class DriverControlled2 extends HardwareDefinitions{
     @Override
     public void runOpMode(){
         init(hardwareMap);
+        initIMU(hardwareMap);
         telemetry.addData("Robot is initialized", "");
 
         waitForStart();
@@ -113,72 +117,15 @@ public class DriverControlled2 extends HardwareDefinitions{
                 landerMotor2.setPower(0);
             }
 
-            if(gamepad1.b){
-                moveBoxMechanism(3,3);
-            }
+            gyroHeading = imu.getAngularOrientation(AxesReference.INTRINSIC, AxesOrder.ZYX, AngleUnit.DEGREES).firstAngle;
+            telemetry.addData("Gyro heading:", gyroHeading);
 
-            if(gamepad1.dpad_right){
-                dropFromLander();
-            }
+
+
         }
         stopRobot();
         telemetry.addData("Robot is stopped", "" );
     }
 
-    public void dropFromLander(){
 
-
-        landerMotor1.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-        landerMotor2.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-
-        int counts_per_rotation = 288; //using Core Hex Motors
-        double rotationsUp = 3.63; //modify this one, will most likely be the same as rotationsDown
-
-        int targetPositionUp1 = landerMotor1.getCurrentPosition() + (int)(counts_per_rotation * rotationsUp);
-        int targetPositionUp2 = landerMotor2.getCurrentPosition() + (int)(counts_per_rotation * rotationsUp);
-
-        landerMotor1.setTargetPosition(targetPositionUp1);
-        landerMotor2.setTargetPosition(targetPositionUp2);
-
-        landerMotor1.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-        landerMotor2.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-
-        //raise the lift and undo the passive latch
-        landerMotor1.setPower(0.4);
-        landerMotor2.setPower(0.4);
-
-        while(opModeIsActive() && (landerMotor1.isBusy() || landerMotor2.isBusy())){
-            //let the motors keep spinning
-        }
-
-        landerMotor1.setPower(0);
-        landerMotor2.setPower(0);
-
-        encoderTurn(0.4, 45, true, 5);
-
-        //bring the lift back down
-        double rotationsDown = 3.63; //modify this one
-
-        int targetPositionDown1 = landerMotor1.getCurrentPosition() - (int)(counts_per_rotation * rotationsDown);
-        int targetPositionDown2 = landerMotor2.getCurrentPosition() - (int)(counts_per_rotation * rotationsDown);
-
-
-        landerMotor1.setTargetPosition(targetPositionDown1);
-        landerMotor2.setTargetPosition(targetPositionDown2);
-
-        landerMotor1.setPower(0.4);
-        landerMotor2.setPower(0.4);
-
-        while(opModeIsActive() && (landerMotor1.isBusy() || landerMotor2.isBusy())){
-            //let the motors keep spinning
-        }
-
-        landerMotor1.setPower(0);
-        landerMotor2.setPower(0);
-
-        encoderTurn(0.4, 45, false, 5);
-
-        telemetry.addData("Landing sequence:", "Complete");
-
-    }
 }
